moved benches from inside pnet_packet/src/ethernet.rs to pnet_packet/benches/packet_benchmarks.rs; added inlining to pnet_macros to increase speed of parsing data  into packet
