+++ pnet_macros_support/src/packet.rs
@@ -64,6 +64,7 @@ macro_rules! impl_index {
         impl<'p> Index<$index_t> for $t<'p> {
             type Output = $output_t;
 
+            #[inline]
             fn index(&self, index: $index_t) -> &$output_t {
                 &self.as_slice().index(index)
             }
@@ -74,6 +75,7 @@ macro_rules! impl_index {
 macro_rules! impl_index_mut {
     ($t:ident, $index_t:ty, $output_t:ty) => {
         impl<'p> IndexMut<$index_t> for $t<'p> {
+            #[inline]
             fn index_mut(&mut self, index: $index_t) -> &mut $output_t {
                 self.as_mut_slice().index_mut(index)
             }
@@ -92,6 +94,7 @@ pub enum PacketData<'p> {
 
 impl<'p> PacketData<'p> {
     /// Get a slice of the packet data.
+    #[inline]
     pub fn as_slice(&self) -> &[u8] {
         match self {
             &PacketData::Owned(ref data) => data.deref(),
@@ -100,11 +103,13 @@ impl<'p> PacketData<'p> {
     }
 
     /// No-op - returns `self`.
+    #[inline]
     pub fn to_immutable(self) -> PacketData<'p> {
         self
     }
 
     /// A length of the packet data.
+    #[inline]
     pub fn len(&self) -> usize {
         self.as_slice().len()
     }
@@ -127,6 +132,7 @@ pub enum MutPacketData<'p> {
 
 impl<'p> MutPacketData<'p> {
     /// Get packet data as a slice.
+    #[inline]
     pub fn as_slice(&self) -> &[u8] {
         match self {
             &MutPacketData::Owned(ref data) => data.deref(),
@@ -135,6 +141,7 @@ impl<'p> MutPacketData<'p> {
     }
 
     /// Get packet data as a mutable slice.
+    #[inline]
     pub fn as_mut_slice(&mut self) -> &mut [u8] {
         match self {
             &mut MutPacketData::Owned(ref mut data) => data.deref_mut(),
@@ -143,6 +150,7 @@ impl<'p> MutPacketData<'p> {
     }
 
     /// Get an immutable version of packet data.
+    #[inline]
     pub fn to_immutable(self) -> PacketData<'p> {
         match self {
             MutPacketData::Owned(data) => PacketData::Owned(data),
@@ -151,6 +159,7 @@ impl<'p> MutPacketData<'p> {
     }
 
     /// Get a length of data in the packet.
+    #[inline]
     pub fn len(&self) -> usize {
         self.as_slice().len()
     }
@@ -179,6 +188,7 @@ pub trait PrimitiveValues {
 
 impl PrimitiveValues for pnet_base::MacAddr {
     type T = (u8, u8, u8, u8, u8, u8);
+    #[inline]
     fn to_primitive_values(&self) -> (u8, u8, u8, u8, u8, u8) {
         (self.0, self.1, self.2, self.3, self.4, self.5)
     }
@@ -186,6 +196,7 @@ impl PrimitiveValues for pnet_base::MacAddr {
 
 impl PrimitiveValues for ::std::net::Ipv4Addr {
     type T = (u8, u8, u8, u8);
+    #[inline]
     fn to_primitive_values(&self) -> (u8, u8, u8, u8) {
         let octets = self.octets();
 
@@ -195,6 +206,7 @@ impl PrimitiveValues for ::std::net::Ipv4Addr {
 
 impl PrimitiveValues for ::std::net::Ipv6Addr {
     type T = (u16, u16, u16, u16, u16, u16, u16, u16);
+    #[inline]
     fn to_primitive_values(&self) -> (u16, u16, u16, u16, u16, u16, u16, u16) {
         let segments = self.segments();
